// src/api/__generated__/story/model/mintActorNftRequestPayMethod.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

export type MintActorNftRequestPayMethod =
  (typeof MintActorNftRequestPayMethod)[keyof typeof MintActorNftRequestPayMethod];

export const MintActorNftRequestPayMethod = {
  point: 'point',
  usdt: 'usdt',
  usdc: 'usdc',
} as const;